Plan 9 from Bell Labs 1997-01-17

+++ carrera/main.c
@@ -402,7 +402,7 @@ userinit(void)
 }
 
 void
-exit(long type)
+exit(int type)
 {
 	uchar *vec;
